+++ database/migrations/2020_01_01_600000_add_page_id_to_pages_table.php
@@ -14,7 +14,7 @@ class AddPageIdToPagesTable extends Migration
     public function up()
     {
         Schema::table('pages', function (Blueprint $table) {
-            $table->unsignedInteger('parent_id')->nullable();
+            $table->unsignedBigInteger('parent_id')->nullable();
             $table->foreign('parent_id')->references('id')->on('pages');
         });
     }
